+++ os/src/sync/mod.rs
@@ -1,9 +1,7 @@
-mod condvar;
 mod mutex;
 mod semaphore;
 mod up;
 
-pub use condvar::Condvar;
 pub use mutex::{FUTEX_WAIT, FUTEX_WAKE};
 pub use mutex::{Mutex, MutexSpin, FutexQ};
 pub use semaphore::Semaphore;
