+++ os/src/syscall/sync.rs
@@ -1,4 +1,4 @@
-use crate::sync::{Condvar, Mutex, Semaphore, FutexQ};
+use crate::sync::{Mutex, Semaphore, FutexQ};
 use crate::task::{TaskStatus, TaskContext};
 use crate::task::{block_current_and_run_next, current_process, current_task, schedule, take_current_task, wakeup_task};
 use crate::timer::{add_timer, get_time_ms};
@@ -49,46 +49,6 @@ pub fn sys_semaphore_down(sem_id: usize) -> isize {
     let sem = Arc::clone(process_inner.semaphore_list[sem_id].as_ref().unwrap());
     drop(process_inner);
     sem.down();
-    0
-}
-
-pub fn sys_condvar_create() -> isize {
-    let process = current_process();
-    let mut process_inner = process.inner_exclusive_access();
-    let id = if let Some(id) = process_inner
-        .condvar_list
-        .iter()
-        .enumerate()
-        .find(|(_, item)| item.is_none())
-        .map(|(id, _)| id)
-    {
-        process_inner.condvar_list[id] = Some(Arc::new(Condvar::new()));
-        id
-    } else {
-        process_inner
-            .condvar_list
-            .push(Some(Arc::new(Condvar::new())));
-        process_inner.condvar_list.len() - 1
-    };
-    id as isize
-}
-
-pub fn sys_condvar_signal(condvar_id: usize) -> isize {
-    let process = current_process();
-    let process_inner = process.inner_exclusive_access();
-    let condvar = Arc::clone(process_inner.condvar_list[condvar_id].as_ref().unwrap());
-    drop(process_inner);
-    condvar.signal();
-    0
-}
-
-pub fn sys_condvar_wait(condvar_id: usize, mutex_id: usize) -> isize {
-    let process = current_process();
-    let process_inner = process.inner_exclusive_access();
-    let condvar = Arc::clone(process_inner.condvar_list[condvar_id].as_ref().unwrap());
-    let mutex = Arc::clone(process_inner.mutex_list[mutex_id].as_ref().unwrap());
-    drop(process_inner);
-    condvar.wait(mutex);
     0
 }
