+++ user/src/sync/mod.rs
@@ -1,4 +1,5 @@
 mod mutex;
+mod condvar;
 
 pub use mutex::*;
